+++ protoBuilds/onsite-ganda-1/pcr_prep1.ot2.apiv2.py.json
@@ -1,0 +1,66 @@
+{
+    "content": "\"\"\"OPENTRONS.\"\"\"\nimport math\n\nmetadata = {\n    'protocolName': 'rhAmpSeq Library Prep Part 1 - PCR Prep 1',\n    'author': 'Opentrons <protocols@opentrons.com>',\n    'source': 'Custom Protocol Request',\n    'apiLevel': '2.11'   # CHECK IF YOUR API LEVEL HERE IS UP TO DATE\n                         # IN SECTION 5.2 OF THE APIV2 \"VERSIONING\"\n}\n\n\ndef run(ctx):\n    \"\"\"PROTOCOL.\"\"\"\n    # [\n    #  num_samples, m20_mount\n    # ] = get_values(  # noqa: F821 (<--- DO NOT REMOVE!)\n    #     \"num_samples\", \"m20_mount\")\n    num_samples, m20_mount = 8, 'right'\n\n\n    # define all custom variables above here with descriptions:\n\n    num_cols = math.ceil(num_samples/8)\n    m20_speed_mod = 4\n    airgap_library = 5\n    # load modules\n    mag_module = ctx.load_module('magnetic module gen2', '1')\n\n    # load labware\n    sample_plate = mag_module.load_labware('nest_96_wellplate'\n                                           '_100ul_pcr_full_skirt')\n    reagent_plate = ctx.load_labware('nest_96_wellplate_100ul_pcr_full_skirt',\n                                     '2')\n    # load tipracks\n    tiprack20 = [ctx.load_labware('opentrons_96_filtertiprack_20ul',\n                                  str(slot))\n                 for slot in [3, 5, 6][:math.ceil(num_samples/32)]]\n\n    # load instrument\n    m20 = ctx.load_instrument('p20_multi_gen2', m20_mount, tip_racks=tiprack20)\n\n    # pipette functions   # INCLUDE ANY BINDING TO CLASS\n\n    # helper functions\n\n    # reagents\n    library_mix = reagent_plate.rows()[0][0]\n    pcr_forward = reagent_plate.rows()[0][1]\n    pcr_reverse = reagent_plate.rows()[0][2]\n    # plate, tube rack maps\n    sample_dest = sample_plate.rows()[0][:num_cols]\n    # protocol\n\n    # add library mix, 5 uL\n    for dest in sample_dest:\n        m20.flow_rate.aspirate /= m20_speed_mod\n        m20.flow_rate.dispense /= m20_speed_mod\n        m20.pick_up_tip()\n        m20.aspirate(5, library_mix)\n        m20.move_to(library_mix.top(-2))\n        ctx.delay(seconds=2)\n        # m20.touch_tip(v_offset=-2)\n        # m20.move_to(library_mix.top(-2))\n        # m20.aspirate(airgap_library, library_mix.top())\n        # m20.dispense(airgap_library, dest.top())\n        m20.dispense(5, dest)\n        m20.mix(1, 5, dest)\n        ctx.max_speeds['A'] = 100\n        ctx.max_speeds['Z'] = 100\n        m20.drop_tip()\n        del ctx.max_speeds['A']\n        del ctx.max_speeds['Z']\n        m20.flow_rate.aspirate *= m20_speed_mod\n        m20.flow_rate.dispense *= m20_speed_mod\n    # add forward, reverse primers, 2 uL each\n    for reagent_source in [pcr_forward, pcr_reverse]:\n        for dest in sample_dest:\n            m20.flow_rate.aspirate /= m20_speed_mod\n            m20.flow_rate.dispense /= m20_speed_mod\n            m20.pick_up_tip()\n            m20.aspirate(2, reagent_source)\n            m20.move_to(reagent_source.top(-2))\n            ctx.delay(seconds=2)\n            # m20.touch_tip(v_offset=-2)\n            m20.move_to(reagent_source.top(-2))\n            # m20.aspirate(airgap_library, reagent_source.top())\n            # m20.dispense(airgap_library, dest.top())\n            m20.dispense(5, dest)\n            m20.mix(1, 5, dest)\n            ctx.max_speeds['A'] = 100\n            ctx.max_speeds['Z'] = 100\n            m20.drop_tip()\n            del ctx.max_speeds['A']\n            del ctx.max_speeds['Z']\n            m20.flow_rate.aspirate *= m20_speed_mod\n            m20.flow_rate.dispense *= m20_speed_mod\n",
+    "custom_labware_defs": [],
+    "fields": [
+        {
+            "default": 96,
+            "label": "Number of Samples",
+            "name": "num_samples",
+            "type": "int"
+        },
+        {
+            "label": "P20 Multi GEN2 mount",
+            "name": "m20_mount",
+            "options": [
+                {
+                    "label": "right",
+                    "value": "right"
+                },
+                {
+                    "label": "left",
+                    "value": "left"
+                }
+            ],
+            "type": "dropDown"
+        }
+    ],
+    "instruments": [
+        {
+            "mount": "right",
+            "name": "p20_multi_gen2"
+        }
+    ],
+    "labware": [
+        {
+            "name": "NEST 96 Well Plate 100 \u00b5L PCR Full Skirt on Magnetic Module GEN2 on 1",
+            "share": false,
+            "slot": "1",
+            "type": "nest_96_wellplate_100ul_pcr_full_skirt"
+        },
+        {
+            "name": "NEST 96 Well Plate 100 \u00b5L PCR Full Skirt on 2",
+            "share": false,
+            "slot": "2",
+            "type": "nest_96_wellplate_100ul_pcr_full_skirt"
+        },
+        {
+            "name": "Opentrons 96 Filter Tip Rack 20 \u00b5L on 3",
+            "share": false,
+            "slot": "3",
+            "type": "opentrons_96_filtertiprack_20ul"
+        },
+        {
+            "name": "Opentrons Fixed Trash on 12",
+            "share": false,
+            "slot": "12",
+            "type": "opentrons_1_trash_1100ml_fixed"
+        }
+    ],
+    "metadata": {
+        "apiLevel": "2.11",
+        "author": "Opentrons <protocols@opentrons.com>",
+        "protocolName": "rhAmpSeq Library Prep Part 1 - PCR Prep 1",
+        "source": "Custom Protocol Request"
+    },
+    "modules": []
+}
